Update belt.json to include Arcane Lanterns
All additions are at the bottom of the list, please double check arcanelanterns:feral_lantern, as it may result in some issues since Arcane Lanterns seems to list it as "minecraft:lantern" as an item, but then lists it as arcanelanterns:feral_lantern as a block. I do not know much about java, so I don't know if this would cause issues.